app/src/test/java/com/xbcoders/xbcad7319/OrderServiceImplTest.kt
package com.xbcoders.xbcad7319

import com.xbcoders.xbcad7319.api.model.Order
import com.xbcoders.xbcad7319.api.service.OrderService
import com.xbcoders.xbcad7319.api.seviceimplementations.OrderServiceImpl
import org.junit.Assert.assertNotNull

import org.junit.Before
import org.junit.Test
import org.mockito.Mock
import org.mockito.Mockito.*
import org.mockito.MockitoAnnotations
import retrofit2.Call

class OrderServiceImplTest {

    private lateinit var orderServiceImpl: OrderServiceImpl

    @Mock
    private lateinit var orderService: OrderService

    @Mock
    private lateinit var placeOrderCall: Call<Order>

    @Mock
    private lateinit var getOrdersCall: Call<List<Order>>

    @Mock
    private lateinit var getUserOrdersCall: Call<List<Order>>

    @Mock
    private lateinit var getOrderByIdCall: Call<Order>

    @Mock
    private lateinit var updateOrderStatusCall: Call<Void>

    @Before
    fun setUp() {
        MockitoAnnotations.openMocks(this)

        // Use reflection to set the orderService in OrderServiceImpl
        // Assuming you have a way to inject the mocked OrderService
        orderServiceImpl = OrderServiceImpl()
        // This step requires that you adjust the OrderServiceImpl to allow setting the orderService mock.
        // Example: orderServiceImpl.setOrderService(orderService)
        // If OrderServiceImpl cannot be modified, consider using a custom test implementation.
    }

    @Test
    fun `placeOrder should call the service to place an order`() {
        // Arrange
        val token = "mocked_token"
        val order = Order() // Create a mock or real Order object as needed
        `when`(orderService.placeOrder(token, order)).thenReturn(placeOrderCall)

        // Act
        val result = orderService.placeOrder(token, order)

        // Assert
        verify(orderService).placeOrder(token, order)
        assertNotNull(result) // Ensure the result matches the expected call
    }

    @Test
    fun `getOrders should call the service to get all orders`() {
        // Arrange
        val token = "mocked_token"
        `when`(orderService.getOrders(token)).thenReturn(getOrdersCall)

        // Act
        val result = orderService.getOrders(token)

        // Assert
        verify(orderService).getOrders(token)
        assertNotNull(result) // Ensure the result matches the expected call
    }

    @Test
    fun `getUserOrders should call the service to get user orders`() {
        // Arrange
        val token = "mocked_token"
        val userId = "mocked_user_id"
        `when`(orderService.getUserOrders(token, userId)).thenReturn(getUserOrdersCall)

        // Act
        val result = orderService.getUserOrders(token, userId)

        // Assert
        verify(orderService).getUserOrders(token, userId)
        assertNotNull(result) // Ensure the result matches the expected call
    }

    @Test
    fun `getOrderById should call the service to get an order by ID`() {
        // Arrange
        val token = "mocked_token"
        val orderId = "mocked_order_id"
        `when`(orderService.getOrderById(token, orderId)).thenReturn(getOrderByIdCall)

        // Act
        val result = orderService.getOrderById(token, orderId)

        // Assert
        verify(orderService).getOrderById(token, orderId)
        assertNotNull(result) // Ensure the result matches the expected call
    }

    @Test
    fun `updateOrderStatus should call the service to update an order status`() {
        // Arrange
        val token = "mocked_token"
        val orderId = "mocked_order_id"
        val status = Order() // Create a mock or real Order object as needed
        `when`(orderService.updateOrderStatus(token, orderId, status)).thenReturn(updateOrderStatusCall)

        // Act
        val result = orderService.updateOrderStatus(token, orderId, status)

        // Assert
        verify(orderService).updateOrderStatus(token, orderId, status)
        assertNotNull(result) // Ensure the result matches the expected call
    }
}